refactor(gtk): Separate state and event handling into their own modules
