+++ gtk/Cargo.toml
@@ -22,3 +22,4 @@ notify-rust = "3.6.2"
 users = "0.9.1"
 upower_dbus = "0.1.0"
 thiserror = "1.0.4"
+derive-new = "0.5.8"
